Adding runtime mesh dispatch for Unstruct Carte

+++ src/model/cartesianSEMmesh/src/test_mesh.cpp
@@ -124,7 +124,7 @@ int main(int argc, char **argv) {
   const int numberOfNodes =
       (size * order + 1) * (size * order + 1) * (size * order + 1);
   // init new mesh
-  CartesianParams<int, float> params{ order, size, size, size, domainSize, domainSize, domainSize };
+  CartesianParams<float, int> params{ order, size, size, size, domainSize, domainSize, domainSize };
   CartesianSEMmesh<float, int, order> mesh(params);
 
   // security checks
